# swagger-spec.json
{"openapi":"3.0.0","paths":{"/users/profile":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredentialsDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/foods":{"post":{"operationId":"FoodsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFoodRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Foods"],"security":[{"bearer":[]}]},"get":{"operationId":"FoodsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Foods"],"security":[{"bearer":[]}]}},"/foods/{id}":{"get":{"operationId":"FoodsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Foods"],"security":[{"bearer":[]}]},"patch":{"operationId":"FoodsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFoodRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["Foods"],"security":[{"bearer":[]}]}},"/recipes":{"post":{"operationId":"RecipesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecipeRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Recipes"],"security":[{"bearer":[]}]},"get":{"operationId":"RecipesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Recipes"],"security":[{"bearer":[]}]}},"/recipes/{id}":{"get":{"operationId":"RecipesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Recipes"],"security":[{"bearer":[]}]},"patch":{"operationId":"RecipesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRecipeRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["Recipes"],"security":[{"bearer":[]}]}},"/portions/{id}":{"patch":{"operationId":"PortionsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePortionRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["Food portions"],"security":[{"bearer":[]}]},"delete":{"operationId":"PortionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Food portions"],"security":[{"bearer":[]}]}},"/db-generator/graphs/init":{"get":{"operationId":"DbGeneratorController_createGraphs","parameters":[],"responses":{"200":{"description":""}},"tags":["Database"]}},"/db-generator/graphs/delete":{"delete":{"operationId":"DbGeneratorController_deleteGraphs","parameters":[],"responses":{"200":{"description":""}},"tags":["Database"]}},"/db-generator/constraints/init":{"get":{"operationId":"DbGeneratorController_createConstraints","parameters":[],"responses":{"200":{"description":""}},"tags":["Database"]}},"/db-generator/constraints/delete":{"delete":{"operationId":"DbGeneratorController_deleteConstraints","parameters":[],"responses":{"200":{"description":""}},"tags":["Database"]}},"/units":{"get":{"operationId":"UnitsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Units"],"security":[{"bearer":[]}]}},"/units/{id}":{"get":{"operationId":"UnitsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Units"],"security":[{"bearer":[]}]}}},"info":{"title":"Identity API","description":"This is the Identity API. Please use this in order to register, authenticate and get information of users.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserRegistrationDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"dateOfBirth":{"format":"date-time","type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["email","password","dateOfBirth","firstName","lastName"]},"UserCredentialsDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateFoodRequestDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"kcalPerKg":{"type":"number","description":"Amount (energy in kcal) of calories per 1 kg of food"},"proteinsPerKg":{"type":"number","description":"Amount (weight in gr) of proteins per 1 kg of food"},"carbohydratesPerKg":{"type":"number","description":"Amount (weight in gr) of carbohydrates per 1 kg of food"},"lipidsPerKg":{"type":"number","description":"Amount (weight in gr) of lipids per 1 kg of food"},"kcalPerLt":{"type":"number","description":"Amount (energy in kcal) of calories per 1 liter of food"},"proteinsPerLt":{"type":"number","description":"Amount (weight in gr) of proteins per 1 litter of food"},"carbohydratesPerLt":{"type":"number","description":"Amount (weight in gr) of carbohydrates per 1 liter of food"},"lipidsPerLt":{"type":"number","description":"Amount (weight in gr) of lipids per 1 liter of food"}},"required":["name","description"]},"UpdateFoodRequestDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"kcalPerKg":{"type":"number","description":"Amount (energy in kcal) of calories per 1 kg of food"},"proteinsPerKg":{"type":"number","description":"Amount (weight in gr) of proteins per 1 kg of food"},"carbohydratesPerKg":{"type":"number","description":"Amount (weight in gr) of carbohydrates per 1 kg of food"},"lipidsPerKg":{"type":"number","description":"Amount (weight in gr) of lipids per 1 kg of food"},"kcalPerLt":{"type":"number","description":"Amount (energy in kcal) of calories per 1 liter of food"},"proteinsPerLt":{"type":"number","description":"Amount (weight in gr) of proteins per 1 litter of food"},"carbohydratesPerLt":{"type":"number","description":"Amount (weight in gr) of carbohydrates per 1 liter of food"},"lipidsPerLt":{"type":"number","description":"Amount (weight in gr) of lipids per 1 liter of food"}}},"CreateFoodPortionRequestDto":{"type":"object","properties":{"foodId":{"type":"string"},"quantity":{"type":"number"},"unitId":{"type":"string"}},"required":["foodId","quantity","unitId"]},"CreateRecipeRequestDto":{"type":"object","properties":{"portions":{"type":"array","items":{"$ref":"#/components/schemas/CreateFoodPortionRequestDto"}},"name":{"type":"string"},"description":{"type":"string"},"realKcalPerKg":{"type":"number","description":"Amount (energy in kcal) of calories per 1 kg of food"},"realWeightInKg":{"type":"number","description":"Measured weight (in kg) after cooking all portions of recipe"}},"required":["portions","name","description"]},"UpdateRecipeRequestDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"realKcalPerKg":{"type":"number","description":"Amount (energy in kcal) of calories per 1 kg of food"},"realWeightInKg":{"type":"number","description":"Measured weight (in kg) after cooking all portions of recipe"}}},"UpdatePortionRequestDto":{"type":"object","properties":{"foodId":{"type":"string"},"quantity":{"type":"number"},"unitId":{"type":"string"}}}}}}